latest from coda-oss

+++ externals/coda-oss/modules/c++/str/unittests/test_str.cpp
@@ -83,9 +83,9 @@ TEST_CASE(testSplit)
 {
     std::string s = "space delimited values are the best!";
     std::vector<std::string> parts = str::split(s, " ");
-    TEST_ASSERT_EQ(parts.size(), 6);
+    TEST_ASSERT_EQ(parts.size(), static_cast<size_t>(6));
     parts = str::split(s, " ", 3);
-    TEST_ASSERT_EQ(parts.size(), 3);
+    TEST_ASSERT_EQ(parts.size(), static_cast<size_t>(3));
     TEST_ASSERT_EQ(parts[2], "values are the best!");
 }
